ml-api/requirements.txt
flask
flask-cors
pymongo
torch
torchvision
Pillow
gunicorn
